Implemented account balance.
Moved all endpoints other than GET to AccountController.
Added account balance field to database.
Tested for some common exception types and validated model method parameters.

+++ Api/Responses/DepositResponse.cs
@@ -1,9 +1,9 @@
 namespace Api.Responses;
 
-public class DepositResponse(int customerId, int accountId, double balance, bool succeeded)
+public class DepositResponse(int customerId, int accountId, decimal balance, bool succeeded)
 {
     public int CustomerId { get; set; } = customerId;
     public int AccountId { get; set; } = accountId;
-    public double Balance { get; set; } = balance;
+    public decimal Balance { get; set; } = balance;
     public bool Succeeded { get; set; } = succeeded;
 }
